City of Austin and Microsoft scripts updated
Updated the salary section for the script to pull jobs from the City of Austin. Also, fixed a bug in the function called generate_summaries for both scripts, so it only generates a new summary when a new job was posted or the job description was updated to save money. Finally, fixed the formatting for the qualifications and job description columns in the final CSV file generated for the Microsoft script.