Updates to Scripts
Updated CityofAustin.py to fix an issue with the salary calculation and the fillna function had to have different inputs as the pandas library is supposedly being updated soon. Updated emerson.py to add new section title names and specify the format of the date when the to_datetime function is called. Finally, updated Job_Board_Scripts_Combined.py to merge the CSV file for each company to one CSV file.